test: rename Mock folder by Dummy
