+++ tests/Unit/Dummy/Customer.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yoanbernabeu\AirtableClientBundle\Tests\Unit\Dummy;
+
+class Customer
+{
+    public int $id;
+    public string $firstName;
+    public string $lastName;
+}
